#221 refactor: 투두 조회 로직과 헤더를 Container 컴포넌트로 통합
- TodoList와 Header를 하나의 Container 컴포넌트로 병합
- 투두 조회 로직을 별도 모듈로 분리
- Context를 제거하고 필요한 상태를 props로 전달하도록 변경
