+++ src/components/Todo/index.ts
@@ -4,6 +4,7 @@ import DeleteMode from '@/components/Todo/DeleteMode';
 import EditMode from '@/components/Todo/EditMode';
 import Status from '@/components/Todo/Status';
 
+export { default as TodoContainer } from '@/components/Todo/Container';
 export { default as TodoForm } from '@/components/Todo/Form';
 export { default as TodoItem } from '@/components/Todo/Item';
 export { default as TodoList } from '@/components/Todo/List';
